+++ R/align_move.R
@@ -24,8 +24,6 @@
 #' @author Jakob Schwalb-Willmann
 #'
 #' @importFrom move timestamps timeLag interpolateTime moveStack move split
-#' @importFrom sp coordinates
-#' @importFrom raster crs
 #' @importFrom lubridate second<- minute<- hour<- day<- 
 #'
 #' @seealso \code{\link{df2move}} \code{\link{frames_spatial}} \code{\link{frames_graph}}
@@ -96,7 +94,7 @@ align_move <- function(m, res = "min", digit = "min", unit = "secs", spaceMethod
     interpolateTime(x, ts.t, spaceMethod)
   }))
   
-  m[,c("x", "y")] <- coordinates(m)
+  m[,c("x", "y")] <- m@coords
   m[,"time"] <- timestamps(m)
   return(m)
 }