+++ tests/testthat/helper-vars.R
@@ -6,6 +6,10 @@ test_maps = as.logical(Sys.getenv("moveVis_test_maps"))
 n_cores <- as.numeric(Sys.getenv("moveVis_n_cores"))
 if(!is.na(n_cores)) if(n_cores > 1) use_multicore(n_cores)
 
+check_mapview <- any(grepl("mapview", installed.packages()[,1]))
+check_leaflet <- any(grepl("leaflet", installed.packages()[,1]))
+
+
 ## which tests to run
 # which_tests = Sys.getenv("moveVis_which_tests")
 # if(which_tests != ""){
@@ -29,6 +33,22 @@ data("basemap_data", package = "moveVis", envir = environment())
 ## movement
 m <- move_data
 m.aligned <- align_move(m, res = 4, unit = "mins")
+
+# shift across dateline
+l.df <- lapply(move::split(m.aligned), as.data.frame)
+df <- do.call(rbind, mapply(x = names(l.df), y = l.df, function(x, y){
+  y$id = x
+  return(y)
+}, SIMPLIFY = F))
+df$x <- df$x+171.06
+df$x[df$x > 180] <- df$x[df$x > 180]-360
+
+m.shifted <- df2move(df, "+proj=longlat +datum=WGS84 +no_defs", "x", "y", "time", "id")
+
+# transform using sf
+df <- sf::st_transform(sf::st_as_sf(m.shifted), sf::st_crs("+init=epsg:3995"))
+df <- cbind.data.frame(sf::st_coordinates(df), time = df$time, id = move::trackId(m.shifted))
+m.shifted.repro <- df2move(df = df, proj = "+init=epsg:3995", x = "X", y = "Y", time = "time", track_id = "id")
 
 ## base map
 r_grad <- basemap_data[[1]]